graphql/requirements.txt
uvicorn~=0.18.3
strawberry-graphql[debug-server]==0.133.5
fastapi==0.85.0
pydantic==1.10.2
SQLAlchemy==1.4.11
../shared_libraries[database]
